Move FML|HS to fml.js
