+++ src/createClient.js
@@ -9,6 +9,7 @@ var yggserver = require('yggdrasil').server({});
 var states = require("./states");
 var debug = require("./debug");
 var UUID = require('uuid-1345');
+var fml = require('./fml');
 
 module.exports=createClient;
 
@@ -53,6 +54,13 @@ function createClient(options) {
   client.once('success', onLogin);
   client.once("compress", onCompressionRequest);
   client.on("set_compression", onCompressionRequest);
+  if(client.forge) {
+    client.on('custom_payload', function(packet) {
+      if (packet.channel === 'FML|HS') {
+        fml.fmlHandshakeStep(client, packet.data);
+      }
+    });
+  }
   if(haveCredentials) {
     // make a request to get the case-correct username before connecting.
     var cb = function(err, session) {
